+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## Unreleased
+
+### Fixed
+- Fixed a bug where empty field layout tabs on Neo block types would be duplicated when saving the Neo field
+
 ## 3.9.2 - 2023-09-23
 
 ### Fixed